Add getTimePassed utility function
